feat: move methods to internal function
