+++ test/recover/ERC20Recover.behavior.js
@@ -1,0 +1,21 @@
+const { expect } = require('chai');
+
+function shouldBehaveLikeERC20Recover(receiver, amount) {
+  context('as a ERC20Recover', function () {
+    describe('_recoverERC20', function () {
+      it('should recover any ERC20', async function () {
+        expect(await this.erc20ToRecover.balanceOf(this.instance.address)).to.be.bignumber.equal(amount);
+        expect(await this.erc20ToRecover.balanceOf(receiver)).to.be.bignumber.equal('0');
+
+        await this.instance.$_recoverERC20(this.erc20ToRecover.address, receiver, amount, { from: receiver });
+
+        expect(await this.erc20ToRecover.balanceOf(this.instance.address)).to.be.bignumber.equal('0');
+        expect(await this.erc20ToRecover.balanceOf(receiver)).to.be.bignumber.equal(amount);
+      });
+    });
+  });
+}
+
+module.exports = {
+  shouldBehaveLikeERC20Recover,
+};
